Repair loading of pam tiles.
Former-commit-id: 252a6127eeb618b79e3d58ba9df76e68f4f67421

+++ console/src/main/java/org/janelia/it/FlyWorkstation/gui/viewer3d/slice_viewer/TileFormat.java
@@ -125,7 +125,7 @@ public class TileFormat
 		this.srgb = srgb;
 	}
 
-	public long getTileBytes() {
+	public int getTileBytes() {
 		int w = getTileSize()[0];
 		int h = getTileSize()[1];
 		int bpp = getChannelCount() * getBitDepth() / 8;
